requirements.txt
django==4.2.6
djangorestframework==3.14.0
rsa==4.9
pyjwt==2.8.0